requirements.txt
Keras==2.4.3
streamlit==0.70.0
imutils==0.5.4
requests==2.22.0
matplotlib==3.1.1
~umpy==1.16.5
opencv_python==4.4.0.40
numpy==1.20.1
Pillow==8.1.2
scikit_learn==0.24.1
